fix: update imports after file rename from types.js to schemas.js
- Fixed all imports from src/lib/ai/types.js to schemas.js
- Fixed TypeScript errors in modelStats optional properties
- Fixed ModelId type usage in message_builder.ts
